refactor(P3-3a): 审批/选择题子系统按状态所有权整体归模块
按计划书判据（共享可变状态的所有权，非行数）切分。实测 handler 表方案需 ~93 项注入面
却只减 1 个共享变量，改切自包含度更高的审批/提问子系统：

- app/approval-questions.js 新增 createApprovalController，接管 activePerm/activeQuestion/
  permExpandBtn/selectedExitMode/multiSelectedIndexes/permIntegrityMismatched/permArmTimer
  七个状态，以及 17 个该子系统独占的 DOM 引用
- app.js 外部只留四个动词：isDuplicate*/enqueue*/resolve*/clearAll/hasPending
- 「清空当前审批与提问」此前在 app.js 里被逐字抄了四遍（epoch 重置/result/error/clearView），
  现收敛为一处 clearAll()
- closeSheet 里的 permArmTimer 泄漏改经 context 桥调 cancelPermArming（避免 TDZ）

app.js 7503→7129 行，顶层 let 96→90。注入面 14 项。E2E 218 全过（含 26 条审批/提问用例）。
